fixed the XID problem.
git-svn-id: https://r-osgi.svn.sourceforge.net/svnroot/r-osgi/trunk@1179 c960cec6-6b13-0410-98af-d8f80e9b129c

+++ transport/mina/src/main/java/ch/ethz/iks/r_osgi/transport/mina/codec/FetchServiceMessageCodec.java
@@ -44,7 +44,7 @@ public class FetchServiceMessageCodec extends RemoteOSGiMessageCodec {
 	public MessageDecoderResult decodeBody(IoSession session, ByteBuffer in,
 			ProtocolDecoderOutput out) throws Exception {
 		final FetchServiceMessage msg = new FetchServiceMessage();
-		msg.setXID(((Short) session.getAttribute("xid")).shortValue());
+		msg.setXID(((Integer) session.getAttribute("xid")).intValue());
 		msg.setServiceID(decodeString(in));
 		out.write(msg);
 		return MessageDecoderResult.OK;
